Resolve tapped room at exact position
No tolerance in tap target (just stubbing in everything else we need)

+++ Eurofurence/Application/Maps/MapsService.swift
@@ -13,8 +13,8 @@ protocol MapsService {
     func add(_ observer: MapsObserver)
     func fetchImagePNGDataForMap(identifier: Map2.Identifier, completionHandler: @escaping (Data) -> Void)
     func fetchContent(for identifier: Map2.Identifier,
-                      atX x: Float,
-                      y: Float,
+                      atX x: Int,
+                      y: Int,
                       completionHandler: @escaping (Map2.Content) -> Void)
 
 }
