+++ EurofurenceTests/Presenter Tests/Maps/Test Doubles/FakeMapsService.swift
@@ -26,15 +26,15 @@ class FakeMapsService: MapsService {
         }
         
         var identifier: Map2.Identifier
-        var x: Float
-        var y: Float
+        var x: Int
+        var y: Int
         var completionHandler: (Map2.Content) -> Void
     }
     
     fileprivate var requests = [ContentRequest]()
     func fetchContent(for identifier: Map2.Identifier,
-                      atX x: Float,
-                      y: Float,
+                      atX x: Int,
+                      y: Int,
                       completionHandler: @escaping (Map2.Content) -> Void) {
         requests.append(ContentRequest(identifier: identifier, x: x, y: y, completionHandler: completionHandler))
     }
@@ -47,7 +47,7 @@ extension FakeMapsService {
         return identifier.rawValue.data(using: .utf8)!
     }
     
-    func resolveMapContents(identifier: Map2.Identifier, atX x: Float, y: Float, with mapContent: Map2.Content) {
+    func resolveMapContents(identifier: Map2.Identifier, atX x: Int, y: Int, with mapContent: Map2.Content) {
         guard let request = requests.first(where: { $0.identifier == identifier && $0.x == x && $0.y == y }) else { return }
         request.completionHandler(mapContent)
     }
